ルーティング変更に際して、フォルダ構成を変更
