+++ acctang/src/app/app-routing.module.ts
@@ -2,17 +2,11 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { PagenotfoundComponent } from './pagenotfound/pagenotfound.component';
-import { ClassificationListComponent } from './classification/classification-list.component';
-import { PurposeListComponent } from './purpose/purpose-list.component';
-import { ExpenseListComponent } from './expense/expense-list.component';
-import { ExpenseInputComponent } from './expense/expense-input.component';
 
 const routes: Routes = [
-  {path: "classification-list", component: ClassificationListComponent},
-  {path: "purpose-list", component: PurposeListComponent},
-  {path: "expense-list", component: ExpenseListComponent},
-  {path: "expense-input", component: ExpenseInputComponent},
-  {path: "**", component: PagenotfoundComponent },
+  { path: 'angular/master', loadChildren: './master/master.module#MasterModule' },
+  { path: 'angular/transaction', loadChildren: './transaction/transaction.module#TransactionModule' },
+  { path: "**", component: PagenotfoundComponent },
 ];
 
 @NgModule({
