+++ acctang/src/app/app.component.html
@@ -1,33 +1,15 @@
-<!--The content below is only a placeholder and can be replaced.-->
 <div style="text-align:center">
   <h1 class="text-center">
     <span class="title">Welcome to {{ title }}</span>
     <p class="sub-title">{{ subtitle }}</p>
   </h1>
-  <img width="300" alt="Angular Logo" src="data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHZpZXdCb3g9IjAgMCAyNTAgMjUwIj4KICAgIDxwYXRoIGZpbGw9IiNERDAwMzEiIGQ9Ik0xMjUgMzBMMzEuOSA2My4ybDE0LjIgMTIzLjFMMTI1IDIzMGw3OC45LTQzLjcgMTQuMi0xMjMuMXoiIC8+CiAgICA8cGF0aCBmaWxsPSIjQzMwMDJGIiBkPSJNMTI1IDMwdjIyLjItLjFWMjMwbDc4LjktNDMuNyAxNC4yLTEyMy4xTDEyNSAzMHoiIC8+CiAgICA8cGF0aCAgZmlsbD0iI0ZGRkZGRiIgZD0iTTEyNSA1Mi4xTDY2LjggMTgyLjZoMjEuN2wxMS43LTI5LjJoNDkuNGwxMS43IDI5LjJIMTgzTDEyNSA1Mi4xem0xNyA4My4zaC0zNGwxNy00MC45IDE3IDQwLjl6IiAvPgogIDwvc3ZnPg==">
 </div>
 
-<button routerLink="/classification-list">分類マスタ一覧</button>
-<button routerLink="/purpose-list">用途マスタ一覧</button>
-<button routerLink="/expense-list">支出一覧</button>
-<button routerLink="/expense-input">支出取込</button>
+<button routerLink="/angular/master">マスタ系画面</button>
+<button routerLink="/angular/transaction">トランザクション系画面</button>
 
 <p>ここから表示</p>
 <br>
 <router-outlet></router-outlet>
 <br>
 <p>ここまで表示</p>
-
-<h2>Here are some links to help you start: </h2>
-<ul>
-  <li>
-    <h2><a target="_blank" rel="noopener" href="https://angular.io/tutorial">Tour of Heroes</a></h2>
-  </li>
-  <li>
-    <h2><a target="_blank" rel="noopener" href="https://angular.io/cli">CLI Documentation</a></h2>
-  </li>
-  <li>
-    <h2><a target="_blank" rel="noopener" href="https://blog.angular.io/">Angular blog</a></h2>
-  </li>
-</ul>
-
