+++ acctang/src/app/master/master-main/master-main.component.html
@@ -1,0 +1,8 @@
+<button routerLink="/classification-list">目的マスタ一覧</button>
+<button routerLink="/purpose-list">用途マスタ一覧</button>
+
+<p>ここから表示</p>
+<br>
+<router-outlet></router-outlet>
+<br>
+<p>mここまで表示</p>
