+++ acctang/src/app/master/master-main/master-main.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+    selector: 'master-main',
+    templateUrl: './master-main.component.html',
+    styleUrls: ['./master-main.component.css']
+})
+export class MasterMainComponent {
+
+    title: string = '＜マスタ系画面＞';
+
+}
+
