+++ acctang/src/app/master/master.module.ts
@@ -1,23 +1,14 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { Injectable, NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
 import { HttpClientModule, HttpClientXsrfModule, HTTP_INTERCEPTORS, HttpInterceptor, HttpXsrfTokenExtractor, HttpRequest, HttpHandler, HttpEvent } from '@angular/common/http';
 
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { PagenotfoundComponent } from './pagenotfound/pagenotfound.component';
+import { MasterMainComponent } from './master-main/master-main.component';
+import { ClassificationListComponent } from '../master/classification-list/classification-list.component';
+import { PurposeListComponent } from '../master/purpose-list/purpose-list.component';
 
-import { ClassificationListComponent } from './master/classification-list/classification-list.component';
-import { PurposeListComponent } from './master/purpose-list/purpose-list.component';
-import { ExpenseListComponent } from './transaction/expense-list/expense-list.component';
-import { ExpenseInputComponent } from './transaction/expense-input/expense-input.component';
-import { IncomeListComponent } from './transaction/income-list/income-list.component';
-
-import { ClassificationService } from './shared/services/classification.service';
-import { PurposeService } from './shared/services/purpose.service';
-import { ExpenseService } from './shared/services/expense.service';
-import { IncomeService } from './shared/services/income.service';
+import { ClassificationService } from '../shared/services/classification.service';
+import { PurposeService } from '../shared/services/purpose.service';
 
 import { Observable } from 'rxjs';
 
@@ -41,14 +32,14 @@ export class HttpXsrfInterceptor implements HttpInterceptor {
 
 @NgModule({
   declarations: [
-    AppComponent,
-    PagenotfoundComponent,
+    MasterMainComponent,
+    ClassificationListComponent,
+    PurposeListComponent,
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpClientModule,
-    AppRoutingModule,
     HttpClientXsrfModule.withOptions({
       cookieName: 'csrftoken',
       headerName: 'X-CSRFToken',
@@ -57,14 +48,11 @@ export class HttpXsrfInterceptor implements HttpInterceptor {
   providers: [
     ClassificationService,
     PurposeService,
-    ExpenseService,
-    IncomeService,
     HttpXsrfInterceptor,
     { provide: HTTP_INTERCEPTORS, useExisting: HttpXsrfInterceptor, multi: true },
   ],
   bootstrap: [
-    AppComponent,
   ]
 })
-export class AppModule { }
+export class MasterModule { }
