+++ acctang/src/app/master/purpose-list/purpose-list.component.css
