+++ acctang/src/app/shared/models/classification.model.ts
@@ -1,4 +1,4 @@
-import { User } from '../user/user.model';
+import { User } from './user.model';
 
 export class Classification {
     id: string;
