+++ acctang/src/app/transaction/transaction-main/transaction-main.component.html
@@ -1,0 +1,8 @@
+<button routerLink="/expense-list">支出一覧</button>
+<button routerLink="/expense-input">支出取込</button>
+
+<p>ここから表示</p>
+<br>
+<router-outlet></router-outlet>
+<br>
+<p>tここまで表示</p>
