+++ acctang/src/app/transaction/transaction-main/transaction-main.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+    selector: 'transaction-main',
+    templateUrl: './transaction-main.component.html',
+    styleUrls: ['./transaction-main.component.css']
+})
+export class TransactionMainComponent {
+
+    title: string = '＜トランザクション系画面＞';
+
+}
+
