+++ acctang/src/app/transaction/transaction.module.ts
@@ -1,23 +1,15 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { Injectable, NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
 import { HttpClientModule, HttpClientXsrfModule, HTTP_INTERCEPTORS, HttpInterceptor, HttpXsrfTokenExtractor, HttpRequest, HttpHandler, HttpEvent } from '@angular/common/http';
 
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { PagenotfoundComponent } from './pagenotfound/pagenotfound.component';
+import { TransactionMainComponent } from './transaction-main/transaction-main.component';
+import { ExpenseListComponent } from '../transaction/expense-list/expense-list.component';
+import { ExpenseInputComponent } from '../transaction/expense-input/expense-input.component';
+import { IncomeListComponent } from '../transaction/income-list/income-list.component';
 
-import { ClassificationListComponent } from './master/classification-list/classification-list.component';
-import { PurposeListComponent } from './master/purpose-list/purpose-list.component';
-import { ExpenseListComponent } from './transaction/expense-list/expense-list.component';
-import { ExpenseInputComponent } from './transaction/expense-input/expense-input.component';
-import { IncomeListComponent } from './transaction/income-list/income-list.component';
-
-import { ClassificationService } from './shared/services/classification.service';
-import { PurposeService } from './shared/services/purpose.service';
-import { ExpenseService } from './shared/services/expense.service';
-import { IncomeService } from './shared/services/income.service';
+import { ExpenseService } from '../shared/services/expense.service';
+import { IncomeService } from '../shared/services/income.service';
 
 import { Observable } from 'rxjs';
 
@@ -41,30 +33,28 @@ export class HttpXsrfInterceptor implements HttpInterceptor {
 
 @NgModule({
   declarations: [
-    AppComponent,
-    PagenotfoundComponent,
+    TransactionMainComponent,
+    ExpenseListComponent,
+    ExpenseInputComponent,
+    IncomeListComponent,
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpClientModule,
-    AppRoutingModule,
     HttpClientXsrfModule.withOptions({
       cookieName: 'csrftoken',
       headerName: 'X-CSRFToken',
     }),
   ],
   providers: [
-    ClassificationService,
-    PurposeService,
     ExpenseService,
     IncomeService,
     HttpXsrfInterceptor,
     { provide: HTTP_INTERCEPTORS, useExisting: HttpXsrfInterceptor, multi: true },
   ],
   bootstrap: [
-    AppComponent,
   ]
 })
-export class AppModule { }
+export class TransactionModule { }
